chore: add types.ts
